[feat] Expand logging engine
Move the primary client logger to the ClientConfig object for
increased internal portability. Add an internal method to log
and error and subsequently throw it, reducing code reuse for this
common pattern. Allow for the easy addition of arbitrary fields to
errors as part of this method, as well as arbitrary (optional) log
messages to accompany the error. General consolidation and improvement
of error handling throughout the library.

This also includes more robust configuration checking with
descritive errors in response to client misconfiguration.

Fixes #3

[feat] Allow auth purging

Add the `authOptions.preserveAuth` configuration parameter.
This defaults to true and replicates the existing behavior of
storing authentication details to allow for the kotne to be
refreshed as needed. Setting this to false causes the credentials
to be dropped by the client after being used for the initial
token request.

[feat] Allow insecure authentication

Add the `authOptions.allowInsecure` configuration parameter.
This defaults to false, which replicates the existing behavior
of refusing to authenticate without a secure channel (SSL/TLS).
Set this parameter to true to enable authentication over
unsecured HTTP.

Fixes #2

[bugfix] Temporarily disable minification

[chore] Update Makefile
[chore] Update unit tests
[chore] Update *ignore files