[feat] Expand logging engine
Move the primary client logger to the ClientConfig object for
increased internal portability. Add an internal method to log
and error and subsequently throw it, reducing code reuse for this
common pattern. Allow for the easy addition of arbitrary fields to
errors as part of this method, as well as arbitrary (optional) log
messages to accompany the error. General consolidation and improvement
of error handling throughout the library.

Released for preliminary testing