[feat] Expand logging engine
Add more robust and stable logging mechanisms. This includes more
helpful error reporting as well as more stringent checking for
configuration parameters. This also applies to insecure auth
(described below) as there are now checks for it and useful
error messages, rather than a cryptic failure.

Fixes #2
Fixes #3

[feat] Allow auth purging

Add the `authOptions.preserveAuth` configuration parameter.
This defaults to true and replicates the existing behavior of
storing authentication details to allow for the kotne to be
refreshed as needed. Setting this to false causes the credentials
to be dropped by the client after being used for the initial
token request.

[feat] Allow insecure authentication

Add the `authOptions.allowInsecure` configuration parameter.
This defaults to false, which replicates the existing behavior
of refusing to authenticate without a secure channel (SSL/TLS).
Set this parameter to true to enable authentication over
unsecured HTTP.

[chore] Update Makefile
[chore] Update unit tests
[chore] Update *ignore files